refactor     extract ObjectSerializer from PrintingConfig
